+++ lib/src/model/auth_session_info.dart
@@ -1,3 +1,5 @@
+import 'dart:convert';
+
 abstract class JSONable {
   Map<String, dynamic> toJson();
 }
@@ -5,6 +7,7 @@ abstract class JSONable {
 abstract class AuthSessionInterface implements JSONable {
   late String accessToken;
   late String phoneNumber;
+  late Map<String, dynamic> others;
 }
 
 class AuthSessionInfo implements AuthSessionInterface {
@@ -12,15 +15,27 @@ class AuthSessionInfo implements AuthSessionInterface {
   late String accessToken;
   @override
   late String phoneNumber;
-  AuthSessionInfo({required this.accessToken, required this.phoneNumber});
+  @override
+  Map<String, dynamic> others;
+  AuthSessionInfo(
+      {required this.accessToken,
+      required this.phoneNumber,
+      this.others = const {}});
 
   factory AuthSessionInfo.fromJson(Map<String, dynamic> json) {
+    Map<String, dynamic>? others = jsonDecode(json['others'] ?? '{}');
     return AuthSessionInfo(
-        accessToken: json['accessToken'], phoneNumber: json['phoneNumber']);
+        accessToken: json['accessToken'],
+        phoneNumber: json['phoneNumber'],
+        others: others ?? {});
   }
 
   @override
   Map<String, dynamic> toJson() {
-    return {'accessToken': accessToken, 'phoneNumber': phoneNumber};
+    return {
+      'accessToken': accessToken,
+      'phoneNumber': phoneNumber,
+      'others': jsonEncode(others)
+    };
   }
 }